feat: 로딩, option selector 공통 컴포저블 정의
